Moved loader function into different action file
Separation of concerns
