+++ ghost-texts/package.json
@@ -25,6 +25,7 @@
     "npm": "^6.1.0",
     "react": "16.3.1",
     "react-native": "~0.55.2",
+    "react-native-animatable": "^1.3.0",
     "react-native-elements": "^0.19.1",
     "react-native-stripe-api": "^0.1.0",
     "react-native-tab-view": "^1.0.2",
